requirements.txt
amqp==5.1.1
asgiref==3.5.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.6
Redis==4.3.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.0.4
django-background-tasks==1.2.5
django-brake==1.6.0
django-compat==1.0.15
django-ratelimit2==0.4.2
django-user-agents==0.4.0
djangorestframework==3.13.1
dnspython==2.2.0
joblib==1.1.0
kombu==5.2.4
mango-jwt==1.3.2
nltk==3.7
nose==1.3.7
passlib==1.7.4
prompt-toolkit==3.0.29
PyJWT==1.7.1
pymongo==3.0.3
django-ratelimit==3.0.1
python-dateutil==2.8.2
pytz==2021.3
regex==2022.4.24
six==1.16.0
sqlparse==0.4.2
tqdm==4.64.0
ua-parser==0.10.0
user-agents==2.2.0
vine==5.0.0
wcwidth==0.2.5
